refactoring and renaming of interpolate functions moved all tests to common module

+++ interpolate/d2v-interpolate-common/src/main/kotlin/io/data2viz/interpolate/rgb.kt
@@ -7,12 +7,13 @@ import io.data2viz.color.colors.rgba
 // TODO must take all types of colors in args (currently RGB only)
 // TODO add alpha interpolation
 // TODO List instead of start, end ? (validate and check size !!)
+// TODO rename interpolate
 fun interpolateRgb(start:Color, end:Color, gamma: Double = 1.0): (Number) -> Color {
     val interpolator = gamma(gamma)
 
-    val r = interpolator(start.r, end.r)
-    val g = interpolator(start.g, end.g)
-    val b = interpolator(start.b, end.b)
+    val r = interpolator(start.r.toDouble(), end.r.toDouble())
+    val g = interpolator(start.g.toDouble(), end.g.toDouble())
+    val b = interpolator(start.b.toDouble(), end.b.toDouble())
 
     return fun(percent: Number) = rgba(r(percent.toDouble()), g(percent.toDouble()), b(percent.toDouble()))
 }
